refactor(ldap): canonicalize DirectoryRow imports
