+++ src/sambatui/ldap/types.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-from collections.abc import Mapping, Sequence
 from dataclasses import dataclass
 from typing import Literal
 
@@ -73,12 +72,3 @@ class LdapServerSettings:
     host: str
     port: int
     use_ssl: bool
-
-
-@dataclass(frozen=True)
-class DirectoryRow:
-    dn: str
-    kind: str
-    name: str
-    summary: str
-    attributes: Mapping[str, Sequence[str]]
