+++ tests/sambatui/smart_views/test_smart_views_edges.py
@@ -7,7 +7,7 @@ from hypothesis import given
 from hypothesis import strategies as st
 
 from sambatui import smart_views as smart_module
-from sambatui.ldap.client import DirectoryRow
+from sambatui.ldap.rows import DirectoryRow
 from sambatui.core.models import DnsRow
 from sambatui.smart_views import (
     ACCOUNTDISABLE,
